Added integer binary operations

+++ aot_refactor/stack.py
@@ -18,7 +18,7 @@ class StackFrame:
     def __getitem__(self, key: str) -> Variable:
         return self.variables[key]
     
-    def allocate(self, name: str, python_type: type, size: MemorySize = MemorySize.DWORD) -> LinesType:
+    def allocate(self, name: str, python_type: type, size: MemorySize = MemorySize.QWORD) -> LinesType:
         self.frame_size += size.value // 8
         self.variables[name] = Variable(name, python_type, OffsetRegister(Reg("rbp"), self.frame_size, True), size)
         return [Ins("sub", Reg("rsp"), size.value // 8)]
@@ -41,7 +41,7 @@ class Stack:
     def current(self) -> StackFrame:
         return self.stack[-1]
     
-    def allocate(self, name: str, python_type: type, size: MemorySize = MemorySize.DWORD) -> LinesType:
+    def allocate(self, name: str, python_type: type, size: MemorySize = MemorySize.QWORD) -> LinesType:
         return self.current.allocate(name, python_type, size)
     
     def push(self):
